fix(storage): align with staging engine and add wrappers
