--- app/game/engine/random.ts
@@ -1,1 +1,0 @@
-export * from '../../_game/engine/random';
