+++ app/game/rules.ts
@@ -1,10 +1,10 @@
-// Wrapper exports for Jest spy-ability; delegates to canonical _game rules
-import * as baseRules from '../_game/rules';
+import * as RULES from '../_game/rules';
 
-export function isValidPlacement(...args: Parameters<typeof baseRules.isValidPlacement>) {
-  return baseRules.isValidPlacement(...args);
-}
+export const isValidPlacement = (
+  board: import('../_game/types').Board,
+  row: number,
+  col: number,
+  value: import('../_game/types').Digit,
+) => RULES.isValidPlacement(board, row, col, value);
 
-export function isSolved(...args: Parameters<typeof baseRules.isSolved>) {
-  return baseRules.isSolved(...args);
-}
+export const isSolved = (board: import('../_game/types').Board) => RULES.isSolved(board);
